Add --with-system-libsafec flag to link against system libsafec
Specifying the --with-system-libsafec flag shall allow the configuration
program to search for and, if found, to link against the libsafec library
that is installed in the system.

After configuring --with-system-libsafec, the compilation will may with
following error:

In file included from /usr/include/libsafec/safe_lib.h:43,
                  from est_server_http.c:39:
/usr/include/libsafec/safe_types.h:42:9: error: unknown type name 'size_t'
   42 | typedef size_t  rsize_t;
      |         ^~~~~~

The system libsafec lacks including stddef.h in its safe_types.h.

Fix that by moving libsafec include directives below the openssl; the latter apparently
does the inclusion of the necessary stddef.h at some point.

Signed-off-by: Aleksandr Makarov <aleksandr.o.makarov@gmail.com>