Fix content-length error on /serverkeyen response
The response did not include the length of the headers for each
part in the response, meaning that the content-length header was
shorter than the content by 245 bytes. This meant that the response
was truncated on clients like Python Requests or curl.
The response was also incorrectly formatted meaning the header for
the private key was not parsed correctly.